+++ src/content/dto/update-content.dto.ts
@@ -8,82 +8,3 @@ export class UpdateContentDto extends PartialType(CreateContentDto) {
   @IsEnum(ContentStatus)
   status?: ContentStatus;
 }
-
-// src/content/dto/content-query.dto.ts
-import {
-  IsOptional,
-  IsEnum,
-  IsString,
-  IsDateString,
-  IsInt,
-  Min,
-  Max,
-  IsUUID,
-  IsArray,
-} from 'class-validator';
-import { Transform } from 'class-transformer';
-import {
-  ContentType,
-  ContentStatus,
-  ContentPriority,
-} from '../enums/content.enums';
-
-export class ContentQueryDto {
-  @IsOptional()
-  @IsInt()
-  @Min(1)
-  @Transform(({ value }) => parseInt(value))
-  page?: number = 1;
-
-  @IsOptional()
-  @IsInt()
-  @Min(1)
-  @Max(100)
-  @Transform(({ value }) => parseInt(value))
-  limit?: number = 10;
-
-  @IsOptional()
-  @IsString()
-  search?: string;
-
-  @IsOptional()
-  @IsEnum(ContentType)
-  type?: ContentType;
-
-  @IsOptional()
-  @IsEnum(ContentStatus)
-  status?: ContentStatus;
-
-  @IsOptional()
-  @IsEnum(ContentPriority)
-  priority?: ContentPriority;
-
-  @IsOptional()
-  @IsUUID()
-  categoryId?: string;
-
-  @IsOptional()
-  @IsArray()
-  @IsUUID('4', { each: true })
-  tagIds?: string[];
-
-  @IsOptional()
-  @IsString()
-  authorId?: string;
-
-  @IsOptional()
-  @IsDateString()
-  publishedAfter?: string;
-
-  @IsOptional()
-  @IsDateString()
-  publishedBefore?: string;
-
-  @IsOptional()
-  @IsString()
-  sortBy?: string = 'createdAt';
-
-  @IsOptional()
-  @IsEnum(['ASC', 'DESC'])
-  sortOrder?: 'ASC' | 'DESC' = 'DESC';
-}
